adding comments and refactoring model
